refactor: split python_plugin.py (1959 LOC) into plugin + extractor
python_extractor.py (new, 1599 lines):
  PythonElementExtractor class — all extraction logic:
  functions, classes, variables, imports, comprehensions, expressions.

python_plugin.py (386 lines, was 1959):
  PythonPlugin class only — thin wrapper, no extraction logic.

No logic changes. Pure file split. Tests updated: patch paths for
extract_text_slice, safe_encode, log_warning changed from
python_plugin → python_extractor module.
