+++ tests/unit/languages/test_python_plugin_edge_cases.py
@@ -115,7 +115,7 @@ class TestPythonPluginEdgeCases:
         extractor.content_lines = ["line1", "line2"]
 
         with patch(
-            "tree_sitter_analyzer.languages.python_plugin.extract_text_slice"
+            "tree_sitter_analyzer.languages.python_extractor.extract_text_slice"
         ) as mock_extract:
             mock_extract.side_effect = Exception("Byte extraction failed")
 
@@ -137,7 +137,7 @@ class TestPythonPluginEdgeCases:
         ]
 
         with patch(
-            "tree_sitter_analyzer.languages.python_plugin.extract_text_slice"
+            "tree_sitter_analyzer.languages.python_extractor.extract_text_slice"
         ) as mock_extract:
             mock_extract.side_effect = Exception("Byte extraction failed")
 
@@ -479,7 +479,7 @@ class TestPythonPluginEdgeCases:
             extractor.content_lines = [f"content_{i}"]
 
             with patch(
-                "tree_sitter_analyzer.languages.python_plugin.extract_text_slice"
+                "tree_sitter_analyzer.languages.python_extractor.extract_text_slice"
             ) as mock_extract:
                 mock_extract.return_value = f"text_{i}"
                 extractor._get_node_text_optimized(mock_node)
@@ -511,7 +511,7 @@ class TestPythonPluginEdgeCases:
 
             # Should handle Unicode without errors
             with patch(
-                "tree_sitter_analyzer.languages.python_plugin.extract_text_slice"
+                "tree_sitter_analyzer.languages.python_extractor.extract_text_slice"
             ) as mock_extract:
                 mock_extract.return_value = unicode_code
                 result = extractor._get_node_text_optimized(mock_node)
@@ -532,7 +532,7 @@ class TestPythonPluginEdgeCases:
 
         # Should handle very long lines
         with patch(
-            "tree_sitter_analyzer.languages.python_plugin.extract_text_slice"
+            "tree_sitter_analyzer.languages.python_extractor.extract_text_slice"
         ) as mock_extract:
             mock_extract.return_value = long_line
             result = extractor._get_node_text_optimized(mock_node)
@@ -557,7 +557,7 @@ class TestPythonPluginEdgeCases:
         mock_node.end_point = (len(extractor.content_lines) - 1, 0)
 
         with patch(
-            "tree_sitter_analyzer.languages.python_plugin.extract_text_slice"
+            "tree_sitter_analyzer.languages.python_extractor.extract_text_slice"
         ) as mock_extract:
             mock_extract.return_value = nested_code
             result = extractor._get_node_text_optimized(mock_node)
@@ -579,7 +579,7 @@ class TestPythonPluginEdgeCases:
 
         # Should handle binary data in strings
         with patch(
-            "tree_sitter_analyzer.languages.python_plugin.extract_text_slice"
+            "tree_sitter_analyzer.languages.python_extractor.extract_text_slice"
         ) as mock_extract:
             mock_extract.return_value = binary_source
             result = extractor._get_node_text_optimized(mock_node)
@@ -626,7 +626,7 @@ class TestPythonPluginEdgeCases:
                 extractor.content_lines = [f"content_{i}"]
 
                 with patch(
-                    "tree_sitter_analyzer.languages.python_plugin.extract_text_slice"
+                    "tree_sitter_analyzer.languages.python_extractor.extract_text_slice"
                 ) as mock_extract:
                     mock_extract.return_value = f"text_{i}"
                     extractor._get_node_text_optimized(mock_node)
@@ -660,7 +660,7 @@ class TestPythonPluginEdgeCases:
 
         # Should fallback gracefully
         with patch(
-            "tree_sitter_analyzer.languages.python_plugin.safe_encode"
+            "tree_sitter_analyzer.languages.python_extractor.safe_encode"
         ) as mock_encode:
             mock_encode.side_effect = Exception("Encoding failed")
